requirements.txt
# Core
streamlit>=1.31.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# Data & Database
duckdb>=0.10.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0  # For anomaly detection

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# LLM & Agent
google-generativeai>=0.8.0
groq>=0.11.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0
python-dateutil>=2.8.0
requests>=2.31.0

# SQL Parsing & Validation
sqlparse>=0.4.4

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# RAG / Embeddings (Enabled for better accuracy)
sentence-transformers>=2.3.0

